Move_PSTs.ps1
Version:  1.4 / Updated:  3/24/2015 / By:  ZT
Change Log:
v1.3 = For some reason, Outlook isn't closing for some users after being
opened the first time.  Added in a timed pause and then stop the process
if it is still running so the script can continue.
v1.4 = Noticed still sometimes attached PSTs were not being moved as
they were "in use" by another process, so this caused new PSTs to be
created with the same name.  Then the function ProfilePSTs then was some
how able to move the original PSTs to the new location, but had to
modify the name since it was now a duplicate.  But the new, empty PSTs
were the ones attached to Outlook after they were created.  I adjusted
the script at the "Add Back to Outlook" section to check to see if the
PST existed in the new location and if not, do not process the "move"
request, so a new PST is not created.  This still allows the function
ProfilePSTs to move the PSTs, but they'll have to be manually added back
to Outlook.  The major point of the script is to move the PST files, and
this doesn't happen often.  The real fix would be to find out what
process is using the files and kill it before the move request (this has
been done with one process).