v1.2 = Reworked check process
v1.1 = Appears the check was not working... Made change.
v1.2 = Realized the check was working, but was not returning the values for the If Statement to compare, so this was causing the values to be inserted on every execution.  Had to use a different method to check if the values are already entered.